fix(build): restore svc_layout link flags, merge fdf_parser into stb_fdf.h
Remove extra comma in call unity_lib_schema that shifted CFLAGS into
the extra-sources slot and dropped all linker flags (-ljass -lshared
etc.), causing the CI and local macOS build to fail with unresolved
symbols.

Delete redundant cl_unit_layout.c — its functions were already merged
into cl_scrn.c in the parent commit, leaving duplicate definitions that
broke the unity build.

Remove stale calls to run_ui_layout_tests/run_ui_logo_layout_tests
whose test files were deleted in commit 51a1cb7.

Merge fdf_parser.c into stb_fdf.h under #ifdef STB_FDF_IMPLEMENTATION
(STB single-header pattern). Add STB_FDF_GLOBALS macro to keep
frames[] as a single definition with -fno-common. Move tokenizer
functions to static inline to avoid symbol conflicts with jlex.c.
Remove static from cross-TU frame API functions.

Add UI_ClearTextureCache() to fix texture-index leak between tests
(ui_texture_names cache not cleared by reset_ui_state).
