+++ games/warcraft-3/tests/stb_fdf_impl.c
@@ -1,0 +1,3 @@
+#define STB_FDF_GLOBALS
+#define STB_FDF_IMPLEMENTATION
+#include "common/stb_fdf.h"
